+++ src/banner/FPSWebPartClass/FPSBaseClass.ts
@@ -84,7 +84,7 @@ export abstract class FPSBaseClass<TProperties> extends BaseClientSideWebPart<TP
   protected _wpInstanceID: string = webpartInstance(this._trickyApp);
   protected _exitPropPaneChanged: boolean = false;
   protected _importErrorMessage: string = '';
-  protected _trickyEmailsALL: string[] = []; // These are emails that get tricky functionality for this specific web part
+  protected _trickyEmailsAll: string[] = []; // These are emails that get tricky functionality for this specific web part
   protected _isSPA: boolean = false; // This is set properly in runSuperOnInit
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any