+++ src/banner/features/PinMe/PinMePropGroup.ts
@@ -3,9 +3,10 @@
 import {
      IPropertyPaneGroup,
      PropertyPaneDropdown, IPropertyPaneDropdownProps,
-      IPropertyPaneDropdownOption,PropertyPaneToggle,
+      IPropertyPaneDropdownOption,PropertyPaneToggle, PropertyPaneLabel,
 
 } from '@microsoft/sp-property-pane';
+import { IThisFPSWebPartClass } from '../../FPSWebPartClass/IThisFPSWebPartClass';
 
 export const PinMeLocations = [
     { index: 0, key: 'normal', text: "normal" },
@@ -14,18 +15,35 @@ export const PinMeLocations = [
     { index: 0, key: 'disabled', text: "disabled" },
   ];
 
-export const FPSPinMePropsGroup : IPropertyPaneGroup = {
-groupName: "Pin Me",
-groupFields: [
-    PropertyPaneDropdown('defPinState', <IPropertyPaneDropdownProps>{
-    label: 'Default Location - "Pin Expanded" updates after save',
-    options: PinMeLocations, //MinHeadingOptions
-    }),
-    //
-    PropertyPaneToggle("forcePinState", {
-    label: "Force Pin State",
-    onText: "Enforce - No toggle",
-    offText: "Let user change",
-    // disabled: true,
-    }),
-]};
+export function FPSPinMePropsGroupX( thisWPClass: IThisFPSWebPartClass ) : IPropertyPaneGroup {
+
+  let fields: any[] = [];
+  if ( thisWPClass._isSPA === true ) {
+    fields.push(
+      PropertyPaneLabel('nothing', {
+        text: 'This set of options is NOT available in SPA Format'
+      }));
+  } else {
+    fields = [
+      PropertyPaneDropdown('defPinState', <IPropertyPaneDropdownProps>{
+      label: 'Default Location - "Pin Expanded" updates after save',
+      options: PinMeLocations, //MinHeadingOptions
+      }),
+      //
+      PropertyPaneToggle("forcePinState", {
+      label: "Force Pin State",
+      onText: "Enforce - No toggle",
+      offText: "Let user change",
+      // disabled: true,
+      }),
+    ]
+  }
+
+  const returnGroup = {
+    groupName: "Pin Me",
+    groupFields: fields 
+  }
+
+  return returnGroup;
+
+}