+++ src/banner/propPane/AllDefaultFPSGroups.ts
@@ -3,7 +3,7 @@ import { FPSEasyPagesGroup } from "../components/EasyPages/EasyPagesGroup";
 import { FPSBanner3VisHelpGroup } from "../components/VisitorPanel/FPSOptionsGroupVisHelp";
 import { FPSOptionsExpando } from "../features/Expando/ExpandoPropGroup";
 import { FPSImportPropsGroup } from "../features/ImportExport/ImportFunctions";
-import { FPSPinMePropsGroup } from "../features/PinMe/PinMePropGroup";
+import { FPSPinMePropsGroupX } from "../features/PinMe/PinMePropGroup";
 import { IThisFPSWebPartClass } from "../FPSWebPartClass/IThisFPSWebPartClass";
 import { FPSBanner3NavGroup } from "./FPSBanner3NavGroup";
 import { FPSBanner3ThemeGroup } from "./FPSBanner3ThemeGroup";
@@ -15,18 +15,20 @@ import { FPSOptionsGroupBasic } from "./FPSOptionsGroupBasic";
  *  With the exception of WebPartInfoGroup which typically is created separately
  * @param thisWPClass 
  */
-export function getAllDefaultFPSFeatureGroups( thisWPClass: IThisFPSWebPartClass ) {
+export function getAllDefaultFPSFeatureGroups( thisWPClass: IThisFPSWebPartClass ): IPropertyPaneGroup[] {
 
   const groups: IPropertyPaneGroup[] = [];
 
-  if ( this._allowPinMe === true ) groups.push( FPSPinMePropsGroup );
-  if ( this._allowEasyPages ) groups.push( FPSEasyPagesGroup( thisWPClass ) );
+  if ( thisWPClass._allowPinMe === true ) groups.push( FPSPinMePropsGroupX( thisWPClass ) );
+  if ( thisWPClass._allowEasyPages ) groups.push( FPSEasyPagesGroup( thisWPClass ) );
   groups.push( FPSBanner3VisHelpGroup( thisWPClass ) );
   groups.push( FPSBanner4BasicGroup( thisWPClass ) );
   groups.push( FPSBanner3NavGroup() );
   groups.push( FPSBanner3ThemeGroup( thisWPClass ) );
-  groups.push( FPSOptionsGroupBasic( thisWPClass ) );
-  if ( this._allowPandoramic ) groups.push( FPSOptionsExpando( thisWPClass ) );
+  if ( thisWPClass._isSPA !== true ) groups.push( FPSOptionsGroupBasic( thisWPClass ) );
+  if ( thisWPClass._allowPandoramic ) groups.push( FPSOptionsExpando( thisWPClass ) );
   groups.push( FPSImportPropsGroup );
-
+  
+  return groups;
+  
 }